--- a/aphrodite/endpoints/openai/protocol.py
+++ b/aphrodite/endpoints/openai/protocol.py
@@ -188,6 +188,9 @@ class ChatCompletionRequest(OpenAIBaseModel):
         default=0,
         validation_alias=AliasChoices("dry_range",
                                       "dry_penalty_last_n"))
+    dry_max_ngram: Optional[int] = 12
+    dry_max_occurrences: Optional[int] = 8
+    dry_early_exit_match_len: Optional[int] = 8
     dynatemp_min: Optional[float] = 0.0
     dynatemp_max: Optional[float] = 0.0
     dynatemp_exponent: Optional[float] = 1.0
@@ -350,6 +353,9 @@ class ChatCompletionRequest(OpenAIBaseModel):
             dry_allowed_length=self.dry_allowed_length,
             dry_sequence_breaker_ids=dry_sequence_breaker_ids,
             dry_range=self.dry_range,
+            dry_max_ngram=self.dry_max_ngram,
+            dry_max_occurrences=self.dry_max_occurrences,
+            dry_early_exit_match_len=self.dry_early_exit_match_len,
             dynatemp_min=self.dynatemp_min,
             dynatemp_max=self.dynatemp_max,
             dynatemp_exponent=self.dynatemp_exponent,
@@ -513,6 +519,9 @@ class CompletionRequest(OpenAIBaseModel):
         default=0,
         validation_alias=AliasChoices("dry_range",
                                       "dry_penalty_last_n"))
+    dry_max_ngram: Optional[int] = 12
+    dry_max_occurrences: Optional[int] = 8
+    dry_early_exit_match_len: Optional[int] = 8
     dynatemp_min: Optional[float] = 0.0
     dynatemp_max: Optional[float] = 0.0
     dynatemp_exponent: Optional[float] = 1.0
@@ -634,6 +643,9 @@ class CompletionRequest(OpenAIBaseModel):
             dry_allowed_length=self.dry_allowed_length,
             dry_sequence_breaker_ids=dry_sequence_breaker_ids,
             dry_range=self.dry_range,
+            dry_max_ngram=self.dry_max_ngram,
+            dry_max_occurrences=self.dry_max_occurrences,
+            dry_early_exit_match_len=self.dry_early_exit_match_len,
             dynatemp_min=self.dynatemp_min,
             dynatemp_max=self.dynatemp_max,
             dynatemp_exponent=self.dynatemp_exponent,