PR:
```
Starting: Batch 1 (20 requests without dry_multiplier)
Batch 1 (20 requests without dry_multiplier): Completed 20 requests in 14.64 seconds
Starting: Batch 2 (10/20 requests with dry_multiplier)
Batch 2 (10/20 requests with dry_multiplier): Completed 20 requests in 19.11 seconds
Starting: Batch 3 (20 requests with dry_multiplier)
Batch 3 (20 requests with dry_multiplier): Completed 20 requests in 23.49 seconds
Total execution time for all batches: 57.24 seconds
```

Main:
```
Starting: Batch 1 (20 requests without dry_multiplier)
Batch 1 (20 requests without dry_multiplier): Completed 20 requests in 14.69 seconds
Starting: Batch 2 (10/20 requests with dry_multiplier)
Batch 2 (10/20 requests with dry_multiplier): Completed 20 requests in 26.89 seconds
Starting: Batch 3 (20 requests with dry_multiplier)
Batch 3 (20 requests with dry_multiplier): Completed 20 requests in 37.49 seconds
Total execution time for all batches: 79.07 seconds
```